pyproject.toml
[tool.poetry]
name = "cmbml"
version = "0.0.0"
description = ""
authors = ["JamesAmato <116766236+JamesAmato@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
torch = "1.12.1"
hydra-core = "^1.3.2"
healpy = "^1.17.1"
tqdm = "^4.66.2"
pysm3 = "^3.4.0"
camb = "^1.5.4"
pandas = "^2.2.2"
scikit-image = "<0.23.2"
scikit-learn = "^1.4.2"
seaborn = "^0.13.2"
cmbnncs = {git = "https://github.com/Guo-Jian-Wang/cmbnncs.git"}
h5py = "^3.11.0"
jinja2 = "^3.1.4"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
pylint = "^3.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"